+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/resources/RegeneratePercentStat.java
@@ -19,7 +19,6 @@ public class RegeneratePercentStat extends Stat {
         this.statRestored = statRestored;
         this.scaling = StatScaling.NONE;
 
-        this.add$To$toTooltip = false;
         this.statGroup = StatGroup.RESTORATION;
 
     }