+++ src/main/java/com/robertx22/age_of_exile/saveclasses/unit/stat_ctx/modify/IStatCtxModifier.java
@@ -7,4 +7,6 @@ public interface IStatCtxModifier {
 
     void modify(ExactStatData thisStat, StatContext target);
 
+    StatContext.StatCtxType getCtxTypeNeeded();
+
 }